requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cloudevents==1.9.0
cryptography==39.0.1
cssselect==1.2.0
DateTime==5.0
deprecation==2.1.0
et-xmlfile==1.1.0
fake-useragent==1.1.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
frozenlist==1.3.3
functions-framework==3.3.0
google-api-core==2.11.0
google-auth==2.16.1
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-oauth==1.0.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
nest-asyncio==1.5.1
numpy==1.24.2
openai==0.10.5
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
pandas-stubs==1.5.3.230214
parse==1.19.0
proto-plus==1.22.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyee==8.2.2
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
requests-html==0.10.0
rsa==4.9
six==1.16.0
soupsieve==2.4
tqdm==4.64.1
types-pytz==2022.7.1.2
urllib3==1.26.14
w3lib==2.1.1
watchdog==2.3.0
websockets==10.4
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
zope.interface==5.5.2
